Correct false positives in CliTest
Two of the tests were passing but for the wrong reasons... the config
file options that were meant to be tested weren't being properly set
because they were prefixed with "--" which is only for CLI options. The
tests were passing due to the actual values being derived from expected
values incorrectly.