GitKsprFunctionalTest bypasses the CLI
The CLI is currently set up to assume each JVM will process only one
command (due to how logging is handled). This isn't ideal for the tests
that do multiple pushes, so I'm switching to wiring everything up
separately for the tests and bypassing the CLI layer. I may revisit this