Properly handle reordered commits
Updating PRs for commits that have been reordered requires special
handling. There is no way to atomically force push the PR branches AND
update their baseRefs. We have to do one or the other first, and if at
any point a PR's `baseRefName..headRefName` is empty, GitHub will
implicitly close that PR and make it impossible for us to update in the
future. To avoid this we temporarily update the baseRefName of any moved
PR to point to ultimate target ref (i.e. "main") first, then force push
the branches, the update the PRs again..