Implement automatic retry on "second level" GitHub API rate limit
GitHub's APIs are rate limited. There is a documented rate limit for the
GraphQL API and the Rest API, but in addition to this there
are undocumented "second level" rate limits. One in particular I'm
seeing in the functional tests are limits on creating pull requests.
This introduces a wrapper that detects rate limit errors and retries
after a series of delays.