Push pushes revision history refs on force pushes (see below)
When (force) pushing `kspr/<id>`, if that ref already exists, we push
it to `kspr/<id>_NN` where NN is 1 higher than the highest existing
NN in the remote, or `01` to start with. Later the description will
be rendered with links to view history between revisions.

All branch updates are atomic, and the revision history refs aren't
force pushed, so in the case of a race between two instances pushing to
the same IDs, one of them will win and the other should fail, which will
avoid lost updates and should hint to the user that someone else is
working on the same set of commits and perhaps some coordination is
called for.