Canonicalize dates coming from JGit
This is in preparation for a new CLI-backed implementation of GitClient.
I want to make sure both are representing dates the same hence this
canonicalize function. I'm admittedly cargo-culting this a bit as my
knowledge of the java time API is wanting and this isn't a high
priority part of the app for me

commit-id: I2440fbd5